p0rk/requirements.txt
Django==1.5
SQLAlchemy==0.8.1
South==0.7.6
amqp==1.0.12
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.31
celery==3.0.21
django-celery==3.0.17
django-dajax==0.9.2
django-dajaxice==0.5.4.1
flower==0.5.0
kombu==2.5.12
psycopg2==2.5
python-dateutil==2.1
pytz==2013b
redis==2.7.4
requests==1.2.0
six==1.3.0
tornado==3.0.1
wsgiref==0.1.2
